Update and rename README.md to Roadmap.md
The core idea of this project is to initiate a transition to a Basic Income based economy through introducing an initial injection of value assets in the form of a wide scope cryptocurrency coins mass distribution and subsequently developing this into an economic ecosystem where value flows are recycled from top to bottom through Basic Income transfers. The system will allow for countercyclical monetary base management and economic policy setting operations, optimally delivered through algorithmic functions embedded in the code.